implementación librería Volley en modalidades de ensayo, creaciónde BD para obtención de estadísticas

+++ app/src/main/java/cl/blackbirdhq/drivit/helpers/AdminSQLiteAPP.java
@@ -8,16 +8,45 @@ import android.database.sqlite.SQLiteOpenHelper;
 public class AdminSQLiteAPP extends SQLiteOpenHelper {
 
     private static final String DB_NAME = "aplications.db";
-    private static final int DB_VERSION = 1;
+    private static final int DB_VERSION = 2;
 
-    private static final String DB_CREATE_QUESTION = "create table questions (_id integer primary key autoincrement," +
-            "question text not null, image text, categories_id integer not null)";
-    private static final String DB_CREATE_ALTERNATIVE = "create table alternatives (_id integer primary key autoincrement," +
-            "alternative text not null, right integer, questions_id integer )";
-    private static final String DB_CREATE_CATEGORY = "create table categories (_id integer primary key autoincrement," +
-            "name text not null, special integer not null)";
-    private static final String DB_CREATE_TEST = "create table test (_id integer primary key autoincrement," +
-            "questions_id integer, alternatives_id integer, right integer)";
+    private static final String QUESTIONS = "create table questions(" +
+            "_id integer primary key autoincrement, " +
+            "question text not null, " +
+            "image text, " +
+            "categories_id integer not null)";
+
+    private static final String ALTERNATIVES = "create table alternatives (" +
+            "_id integer primary key autoincrement," +
+            " alternative text not null," +
+            " right integer not null," +
+            " questions_id integer not null)";
+
+    private static final String CATEGORIES = "create table categories (" +
+            "_id integer primary key autoincrement," +
+            " name text not null," +
+            "special integer not null)";
+
+    private static final String QUESTIONS_TESTS = "create table questions_tests (" +
+            "_id integer primary key autoincrement, " +
+            "questions_id integer not null, " +
+            "tests_id integer not null, " +
+            "selection integer not null, " +
+            "right integer not null)";
+
+    private static final String TESTS = "create table tests (" +
+            "_id integer primary key autoincrement," +
+            " time integer," +
+            " date text not null," +
+            " achieved integer," +
+            " modality text not null," +
+            " type text not null)";
+
+    private static final String DB_CREATE_TEST = "create table test (" +
+            "_id integer primary key autoincrement," +
+            "questions_id integer, " +
+            "alternatives_id integer," +
+            " right integer)";
 
 
 
@@ -25,27 +54,32 @@ public class AdminSQLiteAPP extends SQLiteOpenHelper {
         super(context,  DB_NAME, null,DB_VERSION);
     }
     @Override
-    public void onCreate(SQLiteDatabase db){
-        db.execSQL(DB_CREATE_QUESTION);
-        db.execSQL(DB_CREATE_ALTERNATIVE);
-        db.execSQL(DB_CREATE_CATEGORY);
+    public void onCreate(SQLiteDatabase db) {
+        db.execSQL(QUESTIONS);
+        db.execSQL(ALTERNATIVES);
+        db.execSQL(CATEGORIES);
+        db.execSQL(QUESTIONS_TESTS);
+        db.execSQL(TESTS);
         db.execSQL(DB_CREATE_TEST);
 
     }
 
     public void reloadDBTest(SQLiteDatabase db){
-        db.execSQL("DROP TABLE IF EXISTS questions");
-        db.execSQL("DROP TABLE IF EXISTS alternatives");
-        db.execSQL("DROP TABLE IF EXISTS test");
-        db.execSQL("DROP TABLE IF EXISTS categories");
-        onCreate(db);
+        db.execSQL("DELETE FROM questions");
+        db.execSQL("DELETE FROM alternatives");
+        db.execSQL("DELETE FROM categories");
+        db.execSQL("DELETE FROM test");
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion){
-        /*db.execSQL("DROP TABLE IF EXISTS questions");
-        db.execSQL("DROP TABLE IF EXISTS alternatives");
-        db.execSQL("DROP TABLE IF EXISTS tests");
-        onCreate(db);*/
+        if(oldVersion < 2){
+            db.execSQL("DROP TABLE IF EXISTS questions");
+            db.execSQL("DROP TABLE IF EXISTS alternatives");
+            db.execSQL("DROP TABLE IF EXISTS test");
+            db.execSQL("DROP TABLE IF EXISTS categories");
+            onCreate(db);
+        }
+
     }
 }
